+++ src/Phan/BlockAnalysisVisitor.php
@@ -2535,6 +2535,13 @@ class BlockAnalysisVisitor extends AnalysisVisitor
             // Step into each catch node and get an
             // updated context for the node
             $catch_context = $this->analyzeAndGetUpdatedContext($catch_context, $node, $catch_node);
+            $catch_stmts_node = $catch_node->children['stmts'];
+            if ($catch_stmts_node instanceof Node && BlockExitStatusChecker::willUnconditionallySkipRemainingStatements($catch_stmts_node)) {
+                // e.g. "} catch (Exception $e) { break; }"
+                if (!BlockExitStatusChecker::willUnconditionallyThrowOrReturn($catch_stmts_node)) {
+                    $this->recordLoopContextForBreakOrContinue($catch_context);
+                }
+            }
             // NOTE: We let ContextMergeVisitor->mergeCatchContext decide if the block exit status is valid.
             $catch_context_list[] = $catch_context;
         }